feat(markdown): remove dev from core.py, move them in exports.py
